ai-agents/categorization/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-tester"
version = "0.1.0"
description = "Prompt testing environment for email categorization"
authors = [{name = "Praktikantenamt AI Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "scikit-learn>=1.3.0",
    "rich>=13.7.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.12.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
]

[project.scripts]
prompt-tester = "prompt_tester.cli:main"

[tool.setuptools]
packages = ["prompt_tester"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py310"
